Mock QoE xApp that sends a constant prediction
Version 0.0.1 of an xApplication based on the Python xApp framework.
This mockup listens for a QoE prediction request (RMR message type
30001) as sent by the QP Driver xApp.  Reacts by sending a fixed QoE
prediction (RMR message type 30002), which should be routed to the
Traffic Steering xApp.

Includes a static RMR routing table with this host name for TS:
service-ricxapp-trafficxapp-rmr.svc.cluster.local.

Issue-ID: RICAPP-107
Signed-off-by: Lott, Christopher (cl778h) <cl778h@att.com>
Change-Id: Ie6a6a851dff87c23a0bcaf23d9936e35d0cb3dd9