Update patch set 5
Patch Set 5:

The container release file needs to be reverted back to the original contents.  The job that is failing (#6) shouldn't be triggered until after the change is merged and that container version exists.  The "logic" is that the change merges, and generates a new image in the repo, and then the container-release file is modified, and committed.  That causes the jenkins job to kick which moves the image from staging to release.  The beef it has right now is that it cannot find the "next" image to move because it hasn't been created.

The process has always seemed strange to me, and I don't like having to add commits to "push" the images round, but it is what it is.

I'm still trying to sort the other issue, but figured I'd pass this along now.

Patch-set: 5
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>