Update patch set 5
Patch Set 5:

I don't really know enough about python to know if this is right, but....

Try adding this to the front of your PYTHONPATH:   .:./qp

I pulled the change and was able to duplicate the errors that the jenkins environment is spitting out. When I added this those errors changed to cant find numpy which makes sense as I don't have that installed in the container.   

That's about all I can come up with.

Patch-set: 5
Reviewer: Gerrit User 1000037 <1000037@84e4e190-e623-4c37-9eda-58b1937c1b53>